services/api-gateway/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
psycopg2-binary==2.9.6
redis==5.0.1
confluent-kafka==2.3.0
prometheus-fastapi-instrumentator==6.0.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pydantic==2.5.0
starlette==0.27.0
httpx==0.25.0
python-jose[cryptography]==3.3.0
